requirements.txt
Flask==2.3.3
tensorflow==2.20.0
Pillow==10.1.0
numpy==1.26.0
